Remove WorldInit trait and move all of it's methods into the World trait
