+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 
 - Bump up [MSRV] to 1.62 for more clever support of [Cargo feature]s and simplified codegen. ([fbd08ec2], [cf055ac0], [todo])
 - Replace `#[derive(WorldInit)]` with `#[derive(World)]` to remove manual `World` trait impl. ([#219], [#217])
+- Remove `WorldInit` trait and move all methods from it into the `World` trait. ([#219])
 
 ### Changed
