+++ README.md
@@ -31,7 +31,7 @@ Implement `World` trait and describe steps:
 ```rust
 use std::time::Duration;
 
-use cucumber::{given, then, when, WorldInit as _};
+use cucumber::{given, then, when, World as _};
 use tokio::time::sleep;
 
 #[derive(Debug, Default, cucumber::World)]
