+++ book/src/output/json.md
@@ -12,7 +12,7 @@ And configuring output to [`writer::Json`]:
 ```rust
 # use std::{fs, io};
 # 
-# use cucumber::WorldInit as _;
+# use cucumber::World as _;
 use cucumber::writer;
 
 # #[derive(Debug, Default, cucumber::World)]
