+++ book/src/writing/background.md
@@ -31,7 +31,7 @@ Feature: Animal feature
 ```rust
 # use std::time::Duration;
 #
-# use cucumber::{given, then, when, World, WorldInit as _};
+# use cucumber::{given, then, when, World};
 # use tokio::time::sleep;
 #
 # #[derive(Debug, Default)]
