+++ book/src/writing/doc_strings.md
@@ -53,7 +53,7 @@ Feature: Animal feature
 
 In a [step] matching function, there’s no need to find this text and match it with a pattern. Instead, it may be accessed via [`Step`] argument:
 ```rust,should_panic
-# use cucumber::{gherkin::Step, given, then, when, World, WorldInit as _};
+# use cucumber::{gherkin::Step, given, then, when, World};
 #
 # #[derive(Debug, Default)]
 # struct Cat {
