+++ book/src/writing/rule.md
@@ -25,7 +25,7 @@ Feature: Animal feature
 ```rust
 # use std::time::Duration;
 #
-# use cucumber::{given, then, when, World, WorldInit as _};
+# use cucumber::{given, then, when, World};
 # use tokio::time::sleep;
 #
 # #[derive(Debug, Default)]
