+++ tests/wait.rs
@@ -1,6 +1,6 @@
 use std::{panic::AssertUnwindSafe, time::Duration};
 
-use cucumber::{cli, given, then, when, Parameter, WorldInit as _};
+use cucumber::{cli, given, then, when, Parameter, World as _};
 use derive_more::{Deref, FromStr};
 use futures::FutureExt as _;
 use tokio::time;
